Implement multi-processing
After trying a couple designs, I settled on Marshal in unix socketpairs
to allow for bidirectional communication.

The implementation takes a few shortcuts, and would need to be
hardened quite a bit, particularly in the child managment part.

But the main goal was to expose the parts of the API that would
challenge multi-processing and distribution, which this did effectively.