Harden the multi process implementation
Handle workers dying for whatever reason and requeue the lost
tests if retries are enabled, or fail them if not.

It could be improved further, notably regarding signal handling.
Handling `Interupt` is always a bit wonky, ideally we'd register
traps and use a self pipe to wake up. But this is good enough™
for now.