Implement distributed queue summary
One notable change from ci-queue is that we no longer
keep tabs on individual stats, but simply record all results.

This uses more space in Redis, but allow summry reporters to be
the same as worker reporters. e.g. it's very easy to make a top 10 slow
tests run either on a worker or on the central reporter.