Implement queue heart beats
Taken from ci-queue. Workers regularly update their test lease
TTL. If the worker dies or get stuck somehow, the lease expire
quickly.

When running tests in the main process, we move the heart beat to
a sidecar process to ensure we're not doing network operations
while a test is being executed, as the test might mock various
network operations.

When running in multi-processing mode, we can do the heartbeat
from the main process without worry.