Only load `*_test.rb` and `test_*.rb`
After trying it on Rails test suite it caused boot failures
because it was loading code that was meant as fixtures
to emulate broken code.

I'm kinda annoyed about this change because wrongly naming
a test file that end up never running is a common gotcha
I wanted to clear.

But this is the pragmatic thing to do at least for now.