Support both positional and keyword messages
Positional messages are a big footgun, it's not uncommon to find
test code such as:

```ruby
  assert "expected", actual
```

The intent is to use `assert_equal`, but it passes as a false
positive so is easy to get wrong.

The initial plan was to not support positional messages, but
experiment with some existing test suite shows that it's a bit
too much of a breaking change, so instead we support both
positional and keyword in a way that open the door to make it
an opt-in configuration for legacy test suites.