Implement test isolation
When testing something that requires to heavily change the global
state, or that for some reason may cause a crash or exit of some sort
you may want to run the test in a subprocess.

Right now we only support it if the Ruby platform supports forking
but it should be possible to do it by spawning a new sub process.