Add a verbose reporter
More useful for things like CI or for debugging a leak between
multiple tests.

Would need improvment though as it doesn't work well with
multi-processing and I haven't yet found a good idea to
make it work.

In the future I'd also like the verbose reporter to be
enabled by default when we detect we're running on a CI
system.