List the 5 slowest test on success
Only if they are 10 times slower than the average test.
Also remove the default metrics of cases/s, it's irrelevant,
instead show average / median / p90, p99 test case duration.

Would be nice to be able to ignore them by tagging them slow,
but that would require serializing tags in the results which
is not ideal.